requirements.txt
chess==1.8.0
pocketsphinx2==0.1.17
